requirements.txt
# Dependências principais
Django>=4.2.0,<5.0.0
Pillow>=9.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-widget-tweaks>=1.5
psycopg2-binary>=2.9.0
gunicorn
python-decouple

# Integração com Google Calendar
google-api-python-client==2.149.0
google-auth==2.34.0
google-auth-oauthlib>=0.4.1
google-auth-httplib2==0.2.0
httplib2==0.22.0

# HTTP cliente
requests>=2.28.0

# API REST (para app Flutter)
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0
